+++ aiida/orm/implementation/sql/backends.py
@@ -8,7 +8,6 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """Generic backend related objects"""
-
 import abc
 import typing
 
@@ -37,7 +36,7 @@ class SqlBackend(typing.Generic[ModelType], backends.Backend):
 
         :param model: the ORM model instance to promote to a backend instance
         :return: the backend entity corresponding to the given model
-        :rtype: :class:`aiida.orm.implementation.BackendEntity`
+        :rtype: :class:`aiida.orm.implementation.entities.BackendEntity`
         """
 
     @abc.abstractmethod
